Combine files to reduce size
Use is-buffer to avoid including Buffer during bundle.
