+++ package.json
@@ -23,9 +23,19 @@
   "main": "lib/index.js",
   "module": "libm/index.js",
   "jsnext:main": "libm/index.js",
+  "browser": {
+    "./lib/tersifyFunction.js": "./lib/tersifyFunction.browser.js",
+    "./lib/tersifyFunction.d.ts": "./lib/tersifyFunction.browser.d.ts",
+    "./lib/tersifyFunction.js.map": "./lib/tersifyFunction.browser.js.map",
+    "./libm/tersifyFunction.js": "./libm/tersifyFunction.browser.js",
+    "./libm/tersifyFunction.d.ts": "./libm/tersifyFunction.browser.d.ts",
+    "./libm/tersifyFunction.js.map": "./libm/tersifyFunction.browser.js.map",
+    "./src/tersifyFunction.ts": "./src/tersifyFunction.browser.ts"
+  },
   "typings": "lib/index.d.ts",
   "scripts": {
     "build": "tsc -p tsconfig.es5.json && tsc -p tsconfig.esnext.json",
+    "bundle": "webpack",
     "clean": "rimraf lib libm",
     "dependency-check": "dependency-check . --unused --no-dev && dependency-check . --missing --no-dev",
     "lint": "tslint -p tsconfig.json",
@@ -35,10 +45,17 @@
   },
   "dependencies": {
     "acorn": "^6.1.1",
-    "acorn-bigint": "^0.3.1",
-    "unpartial": "^0.5.0"
+    "acorn-bigint": "^0.4.0",
+    "is-buffer": "^2.0.3",
+    "unpartial": "^0.5.1"
   },
   "devDependencies": {
-    "@unional/devpkg-node": "^1.2.5"
+    "@unional/devpkg-node": "^1.2.5",
+    "param-case": "^2.1.1",
+    "pascal-case": "^2.0.1",
+    "ts-loader": "^6.0.0",
+    "webpack": "^4.31.0",
+    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack-cli": "^3.3.2"
   }
 }
