--- src/tersifyValue/index.ts
@@ -1,3 +1,0 @@
-export * from './interfaces';
-export * from './tersifyValue';
-export * from './trim';
