--- src/tersifyValue/tersifyString.ts
@@ -1,6 +1,0 @@
-import { TersifyContext } from './interfaces';
-import { trim } from './trim';
-
-export function tersifyString(context: TersifyContext, value: string, length: number) {
-  return trim(context, `'${value}'`, length)
-}
