+++ src/typesInternal.ts
@@ -1,3 +1,3 @@
-import { TersifyOptions } from '../interfaces';
+import { TersifyOptions } from './types';
 
 export type TersifyContext = TersifyOptions & { path: (keyof any)[], noTrim: boolean, references: { value: any, path: (keyof any)[] }[] }
