+++ wallaby.js
@@ -1,15 +1,15 @@
 module.exports = function (wallaby) {
   return {
-    "files": [
+    'files': [
       { pattern: 'tsconfig.*', instrument: false },
-      "src/**/*.ts",
-      "!src/**/*.spec.ts"
+      'src/**/*.ts',
+      '!src/**/*.spec.ts'
     ],
-    "tests": [
-      "src/**/*.spec.ts"
+    'tests': [
+      'src/**/*.spec.ts'
     ],
-    "env": {
-      "type": "node"
+    'env': {
+      'type': 'node'
     },
     compilers: {
       'src/**/*.ts': wallaby.compilers.typeScript({ module: 'commonjs' }),
@@ -18,6 +18,6 @@ module.exports = function (wallaby) {
       allowIgnoringCoverageInTests: true,
       ignoreCoverage: /istanbul ignore next/
     },
-    testFramework: 'ava'
+    testFramework: 'jest'
   }
 }
